Update MutationClassifiers-Copy2.ipynb
This is for text pre-processing; the code takes an Irish word and classifies it as eclipsed, lenited, or neither. Does so by iterating through the word's onset, matching to list membership for lenition, which follows the regular pattern 'Xh...,' and list membership and dictionary value for eclipsis, which has several patterns.  Ensures 'bh...' and 'bhf...' return lenition and eclipsis respectively. Better for text with standard formal orthography; suggestion for operator modification for use on corpora more likely to include non-standard spelling included at relevant line.